requirements/requirements_base.txt
# requirements/requirements_base.txt
# 모든 환경에서 공통으로 필요한 기본 라이브러리들

# 기본 과학 계산 라이브러리
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0

# 이미지 처리 라이브러리
pillow>=8.3.0
opencv-python>=4.5.0

# 유틸리티 라이브러리
tqdm>=4.62.0

# 웹 API 프레임워크
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
jinja2>=3.0.0

# 데이터베이스 및 ORM
sqlalchemy>=1.4.0

# 시각화 라이브러리
plotly>=5.3.0

# 머신러닝 유틸리티
scikit-learn>=1.0.0